Divide BarGroup into its own file
